requirements.txt
accelerate
aiohttp
aiohttp-cors
aiosignal
annotated-types
anyio
async-timeout
attrs
cachetools
certifi
charset-normalizer
click
colorful
git+https://github.com/all-secure-src/apt-diffusers-v190524.git
distlib
exceptiongroup
fastapi
filelock
frozenlist
fsspec
google-api-core
google-auth
googleapis-common-protos
grpcio
h11
httptools
huggingface-hub
idna
importlib_metadata
Jinja2
jsonschema
jsonschema-specifications
linkify-it-py
markdown-it-py
MarkupSafe
mdit-py-plugins
mdurl
memray
mpmath
msgpack
multidict
networkx
numpy
opencensus
opencensus-context
packaging
pandas
pillow
platformdirs
prometheus_client
proto-plus
protobuf
psutil
py-spy
pyarrow
pyasn1
pyasn1_modules
pydantic==1.8
pydantic_core
Pygments
python-dateutil
python-dotenv
python-multipart
pytz
PyYAML
ray
referencing
regex
requests
rich
rpds-py
rsa
safetensors
six
smart-open
sniffio
starlette
sympy
textual
tokenizers
torch
torchaudio
torchvision
tqdm
transformers
triton
typing_extensions
tzdata
uc-micro-py
urllib3
uvicorn
uvloop
virtualenv
watchfiles
websockets
wrapt
xformers
yarl
zipp